requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchainhub
langchain-groq
chromadb
pypdf
python-dotenv
